requirements.txt
python-dotenv==1.0.0
openai==2.6.1